# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Document for API Project",
        "version": "1.0"
    },
    "paths": {
        "/api/": {},
        "api/users/create": {
            "post": {
                "tags": [
                    "User CRUD"
                ],
                "summary": "create user",
                "operationId": "765ae49e5cc7b2a02b2c44d91773b673",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "Alex"
                                            },
                                            "email": {
                                                "type": "string",
                                                "example": "alex@gmail.com"
                                            },
                                            "password": {
                                                "type": "string",
                                                "example": "123456789"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "create done",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Alex"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "alex@gmail.com"
                                        },
                                        "password": {
                                            "type": "string",
                                            "example": "hash passord"
                                        },
                                        "createAdd": {
                                            "type": "integer",
                                            "example": "12.12.2024"
                                        },
                                        "updateAdd": {
                                            "type": "integer",
                                            "example": "12.12.2024"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "User CRUD",
            "description": "User CRUD"
        }
    ]
}